Load/play attack hit sounds; update editor layout
Add audio assets and runtime support for attack hit sounds, plus editor/layout tweaks and minor animation/C++ fixes.

- Added three whoosh WAVs for boss/player feedback.
- Added AttackHit.Sound defaults in PlayerConfig.lua.
- Implemented sound handling in PlayerFeedback.lua: key resolution, sound loading cache, EnsureConfiguredSoundLoaded, PlayConfiguredSound; play AttackHit sound and pre-load at Init; clear cache on Shutdown.
- Updated several Samurai animation .uasset files (binary changes).
- Adjusted editor settings (Editor.ini, imgui.ini) to change content browser path, camera/view settings, viewport speed, UI visibility and dock/layout sizes.
- Fixed type/template usage in AnimationTimelinePanel.cpp (use int32 for std::max/min and sampling/frame step calculations).

These changes enable configurable attack-hit audio feedback and refine editor UI/layout and timeline sampling behavior.

+++ KraftonEngine/Source/Editor/UI/Asset/Animation/AnimationTimelinePanel.cpp
@@ -700,8 +700,8 @@ namespace
 			return 0.0f;
 		}
 
-		const int32 PrevIndex = std::max<float>(0, KeyIndex - 1);
-		const int32 NextIndex = std::min<float>(NumKeys - 1, KeyIndex + 1);
+		const int32 PrevIndex = std::max<int32>(0, KeyIndex - 1);
+		const int32 NextIndex = std::min<int32>(NumKeys - 1, KeyIndex + 1);
 		const FRawFloatCurveKey& Prev = Curve.Keys[PrevIndex];
 		const FRawFloatCurveKey& Next = Curve.Keys[NextIndex];
 		const float DeltaTime = Next.TimeSeconds - Prev.TimeSeconds;
@@ -1132,7 +1132,7 @@ void FAnimationTimelinePanel::Render(UAnimSingleNodeInstance* NodeInst,
 
 	// ── 룰러 눈금 / 프레임 번호 ──
 	const int RawStep = static_cast<int>(std::lround(NumFrames * 55.0f / CanvasW));
-	const int Step    = NiceFrameStep(std::max<float>(RawStep, 1));
+	const int Step    = NiceFrameStep(std::max<int>(RawStep, 1));
 	for (int F = 0; F <= EndFrame; ++F)
 	{
 		const float X = TimeToX((static_cast<float>(F) / EndFrame) * PlayLength);
@@ -1660,7 +1660,7 @@ void FAnimationTimelinePanel::Render(UAnimSingleNodeInstance* NodeInst,
 					{
 						ImVec2 PrevPoint;
 						bool bHasPrevPoint = false;
-						const int32 SampleCount = std::max<float>(32, static_cast<int32>(CanvasW / 8.0f));
+						const int32 SampleCount = std::max<int32>(32, static_cast<int32>(CanvasW / 8.0f));
 						for (int32 Sample = 0; Sample <= SampleCount; ++Sample)
 						{
 							const float T = (static_cast<float>(Sample) / static_cast<float>(SampleCount)) * PlayLength;
